+++ hack/build-test-bin.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-set -e
+# set -e
 set +x
 
 if ! which go; then
@@ -23,7 +23,20 @@ mkdir -p bin
 function build_and_move_suite {
 	suite=$1
 	target=$2
-	ginkgo build ./test/"$suite"
+	origin_url="https://github.com/redhat-eets/sno-tests/"
+	git_tag=$(git describe --tags)
+	if [ $? -ne 0 ]; then
+		git_tag=$(git rev-parse HEAD)
+		if [ $? -ne 0 ]; then
+			origin_url="Not Found"
+		else
+			origin_url="${origin_url}tree/${git_tag}"
+		fi
+	else
+		git_tag=$(echo $git_tag | sort -V | tail -1)
+		origin_url="${origin_url}releases/tag/${git_tag}"
+	fi
+	ginkgo build -ldflags "-X 'github.com/redhat-eets/sno-tests/test/ptp.origin_url=${origin_url}'" ./test/"$suite"
 	mv ./test/"$suite"/"$suite".test "$target"
 }
